Add files via upload
This assignment required my team to create a dashboard with ArcGis (interactive web maps) charts and customizations of our own in PowerBI to show the state of the COVID-19 outbreak at the time. I have included charts for the world and for India specifically which was the requirement of the assignment.